Adding UserProfileContoller, Editing User Profile Repo to Long

+++ FRS/src/main/java/com/group/FRS/repository/User_ProfileRepository.java
@@ -7,6 +7,6 @@ import com.group.FRS.model.UserProfile;
 
 
 @Repository
-public interface User_ProfileRepository extends JpaRepository<UserProfile, Integer> {
+public interface User_ProfileRepository extends JpaRepository<UserProfile, Long> {
 	//
 	}
